Moved Tax Details to BillHolder
