+++ app/src/main/java/com/checkin/app/checkin/session/scheduled/ScheduledSessionCartView.kt
@@ -7,7 +7,10 @@ import android.util.AttributeSet
 import android.view.MotionEvent
 import android.view.View
 import android.view.ViewGroup
-import android.widget.*
+import android.widget.EditText
+import android.widget.FrameLayout
+import android.widget.ImageView
+import android.widget.TextView
 import androidx.core.content.ContextCompat
 import androidx.core.widget.NestedScrollView
 import androidx.fragment.app.FragmentActivity
@@ -30,16 +33,11 @@ import com.checkin.app.checkin.misc.BillHolder
 import com.checkin.app.checkin.misc.BlockingNetworkViewModel
 import com.checkin.app.checkin.session.models.ScheduledSessionDetailModel
 import com.checkin.app.checkin.session.models.SessionPromoModel
-import com.checkin.app.checkin.session.models.TaxDetailModel
 import com.checkin.app.checkin.session.scheduled.viewmodels.NewScheduledSessionViewModel
 import com.checkin.app.checkin.utility.LockableBottomSheetBehavior
 import com.checkin.app.checkin.utility.Utils
 import com.checkin.app.checkin.utility.pass
 import com.google.android.material.bottomsheet.BottomSheetBehavior
-import com.skydoves.balloon.ArrowOrientation
-import com.skydoves.balloon.BalloonAnimation
-import com.skydoves.balloon.createBalloon
-import kotlinx.android.synthetic.main.tooltip_tax_details.view.*
 
 class ScheduledSessionCartView @JvmOverloads constructor(
         context: Context, attrs: AttributeSet? = null, defStyleAttr: Int = 0
@@ -80,8 +78,6 @@ class ScheduledSessionCartView @JvmOverloads constructor(
     internal lateinit var tvRestaurantName: TextView
     @BindView(R.id.tv_cart_header_restaurant_locality)
     internal lateinit var tvRestaurantLocality: TextView
-    @BindView(R.id.container_invoice_tax_details)
-    internal lateinit var llTaxDetails: LinearLayout
 
     lateinit var activity: FragmentActivity
     lateinit var viewModel: ScheduledCartViewModel
@@ -92,23 +88,6 @@ class ScheduledSessionCartView @JvmOverloads constructor(
     private val ordersController = CartOrderedItemController(this)
     private val billHolder: BillHolder
     private lateinit var bottomSheetBehavior: LockableBottomSheetBehavior<View>
-
-    private val balloonTaxDetails by lazy {
-        createBalloon(context) {
-            layout = R.layout.tooltip_tax_details
-            arrowOrientation = ArrowOrientation.BOTTOM
-            backgroundColor = ContextCompat.getColor(context, R.color.pinkish_grey)
-            balloonAnimation = BalloonAnimation.FADE
-            padding = 1
-            setBackgroundDrawableResource(R.drawable.layout_border_grey)
-            lifecycleOwner = activity
-            dismissWhenClicked = true
-            dismissWhenShowAgain = true
-            dismissWhenTouchOutside = true
-
-
-        }
-    }
 
     init {
         View.inflate(context, R.layout.fragment_scheduled_session_cart, this).apply {
@@ -131,14 +110,6 @@ class ScheduledSessionCartView @JvmOverloads constructor(
         }
         epoxyRvCartOrders.setHasFixedSize(false)
         epoxyRvCartOrders.setControllerAndBuildModels(ordersController)
-
-        llTaxDetails.setOnClickListener {
-            if (balloonTaxDetails.isShowing) {
-                balloonTaxDetails.dismiss()
-            } else {
-                balloonTaxDetails.showAlignTop(it)
-            }
-        }
     }
 
     fun isExpanded() = bottomSheetBehavior.state == BottomSheetBehavior.STATE_EXPANDED
@@ -146,9 +117,6 @@ class ScheduledSessionCartView @JvmOverloads constructor(
     fun dismiss() {
         if (isExpanded()) {
             bottomSheetBehavior.state = BottomSheetBehavior.STATE_COLLAPSED
-            if (balloonTaxDetails.isShowing) {
-                balloonTaxDetails.dismiss()
-            }
         }
     }
 
@@ -238,7 +206,6 @@ class ScheduledSessionCartView @JvmOverloads constructor(
                 networkViewModel.updateStatus(resource, LOAD_DATA_CART_DETAILS)
                 if (resource.status == Resource.Status.SUCCESS && resource.data != null) {
                     setupData(resource.data)
-                    setupTaxDetails(resource.data.bill.taxDetail)
                 }
             }
         })
@@ -348,18 +315,6 @@ class ScheduledSessionCartView @JvmOverloads constructor(
     }
 
     override fun onTextChanged(s: CharSequence?, start: Int, before: Int, count: Int) {
-    }
-
-    private fun setupTaxDetails(taxDetailModel: TaxDetailModel) {
-        balloonTaxDetails.getContentView().apply {
-            tv_tax_cgst.text = "₹ ${taxDetailModel.cgst}"
-            tv_tax_sgst.text = "₹ ${taxDetailModel.sgst}"
-            if (taxDetailModel.igst.toFloat() != 0F) {
-                tv_tax_igst.visibility = View.VISIBLE
-                tv_tax_igst_heading.visibility = View.VISIBLE
-                tv_tax_igst.text = "₹ ${taxDetailModel.igst}"
-            }
-        }
     }
 
     companion object {
